trace-metric-demo/function_src/requirements.txt
google-cloud-monitoring==0.34.0
googleapis-common-protos==1.51.0
google-cloud-firestore==1.6.1